refactor(scenarios): extract shared file-drop handler into common (TH-7036)
