+++ frontend/src/sections/scenarios/common.js
@@ -262,3 +262,44 @@ export const extractVersionFromScenarioName = (name, basePattern) => {
   const match = name.match(regex);
   return match ? parseInt(match[1], 10) : 0;
 };
+
+export const createScenarioFileDropHandler =
+  ({ enqueueSnackbar, onChange }) =>
+  (acceptedFiles, fileRejections = []) => {
+    if (fileRejections.length > 0) {
+      const hasTypeError = fileRejections.some((rejection) =>
+        rejection.errors?.some((err) => err.code === "file-invalid-type"),
+      );
+      enqueueSnackbar(
+        hasTypeError
+          ? "Unsupported file type. Please upload a TXT or PDF file."
+          : "File could not be uploaded",
+        { variant: "error" },
+      );
+      return;
+    }
+
+    const files = Array.from(acceptedFiles);
+    const maxSize = 5 * 1024 * 1024; // 5MB
+
+    const filesLargerThanMaxSize = files.filter((file) => file?.size > maxSize);
+
+    if (filesLargerThanMaxSize.length > 0) {
+      enqueueSnackbar("File size is too large", {
+        variant: "error",
+      });
+      return;
+    }
+
+    const validFiles = files.filter((file) => file.size <= maxSize);
+
+    const processedFiles = validFiles.map((file) => ({
+      file: file,
+      name: file.name,
+      size: file.size,
+    }));
+
+    if (onChange) {
+      onChange(processedFiles?.[0]);
+    }
+  };
